changed tsx files to js and commented out any ts code
