+++ package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "@emotion/react": "^11.13.5",
     "@emotion/styled": "^11.13.5",
+    "@mantine/core": "^7.14.3",
     "@mui/icons-material": "^6.1.9",
     "@mui/material": "^6.1.9",
     "react": "^18.3.1",
