+++ timesheet-client/src/index.js
@@ -5,8 +5,9 @@ import './index.css';
 import App from './App';
 
 const root = ReactDOM.createRoot(
-  document.getElementById('root') as HTMLElement
+  document.getElementById('root')
 );
+
 root.render(
   <React.StrictMode>
     <App />
